codec: refactor to afford bounds check elimination
See https://docs.google.com/document/d/1vdAEAjYdzjnPA9WDOQ1e4e05cYVMpqSxJYZT33Cqw2g

- use uints for slicing and indexing,
  as it eliminates if index <= 0 checks
- use for j = 0; j < len(x); j++ pattern
  which is identified by golang's bounds check elimination prove phase
- Try to use constants within slices (instead of computing the index).
  This may involve making temporary slices, and working off that.
- Instead of json's readLit (for null, true, false),
  create dedicated functions that use constants to compare.
- eliminate bitset32 and bitset128, which do not elide
  bounds checks, because the pos (byte) can range from 0..255.
  Instead, use bitset256 where the pos is always in range, and bounds checks
  are eliminated.
- Use arithmetic whose value must be in the range of an array
  e.g. for immutableKinds which is [32]reflect.Kind, take the parameter%32.
  Compiler knows that the value is in range 0..31, so not bounds checks

+++ codec/binc.go
@@ -534,9 +534,9 @@ func (d *bincDecDriver) decUint() (v uint64) {
 		d.r.readb(d.b[4:8])
 		v = uint64(bigen.Uint32(d.b[4:8]))
 	case 4, 5, 6:
-		lim := int(7 - d.vs)
+		lim := 7 - d.vs
 		d.r.readb(d.b[lim:8])
-		for i := 0; i < lim; i++ {
+		for i := uint8(0); i < lim; i++ {
 			d.b[i] = 0
 		}
 		v = uint64(bigen.Uint64(d.b[:8]))
